RoleAppointments should not be able to overlap.
We want to be able to add historic appointments so we can associate speeches with ministers in their previous roles. However, we don't yet know whether or not it is meaningful for the same role to be held by more than one person at more than one time, so we've decided that the application should validate against that until such point that we are certain it's something we need to model (at which point things like Role#current_person would need to change to Role#current_people anyway).

We've implemented an amazing ASCIIART-test mechanism (patent pending David Heath) to describe the test fixtures for appointments that should and should not be considered as overlapping.

We've preserved the existing behaviour whereby an new appointment created via the web interface always makes itself current and closes any current appointment.