Don't allow new appointments to be created that will overlap.
Previously, it was possible to create a new appointment that started before
the current one, which would result in the existing appointment having an
`ended_at` before its `started_at`. Now validation prevents this from
happening.

In order to make the error clearer, I've also presented the current
appointments when creating or editing appointments.