`Role#to_s` shouldn't know anything about the current person. Using a default `to_s` method from `Role` everywhere that a minister was used was, I believe, muddying the distinction between Role, RoleAppointment and Person, and made it harder to sensibly construct strings and links to each which included just the right amount of context to be useful.
I have extracted the `current_person` usage out of the Role (except for the search_index, although perhaps that's a pointer that we should really be indexing appointments, or people, or some other combination of information?).

N.B. This was mostly written by James, and then merged in by Tom, hence the pairing.