Allow policies to be associated with Policy Teams
There is a new dropdown list on the admin policy page that allows a user
to associate a Policy Team with the Policy they're working on.  Some
oddness in the chosen jquery plugin means that you can't currently
unassign a Policy Team, although you can change it to a different team.

The Policy Team is shown on the public facing Policy and Supporting Page
pages and in the admin policy view.