Display organisation names in ministerial roles select box.
The implicit call to `Organisation#to_s` appears to be an unintended
remnant of the changes made in b0e99e3163e47bcaaf37459b8d3752a17747e88d.

I've added a test to highlight the problem. However the test duplicates
a lot of the setup in another test. I intend to remove this duplication
in a later commit.

Although I can see that using implicit #to_s methods is convenient and
makes for terser code, I do suspect that using a separate explicit
method would make the code more maintainable.