Make policy team admin more consistent.
We typically have an entity's name be a link to the edit page for that
entity. Originally a policy team only had an email address and not a
name which is why the email address was previously the link the admin
page. This is particularly confusing, because elsewhere in the admin
interface, the email address is a "mailto" link.