Allow specialist guides to be associated with topics
I refactored the admin controller tests covering the association of
policies with topics into a general purpose
should_allow_association_with_topics helper, shared with specialist
guides controller test.

We're expecting to do further work on the layout of the frontend topics
page, so this layout is very 'draft' for now.