Update chosen.js to v0.9.8 (by tag).
There are later commits, but I decided to used the tagged version
because it included the fix I wanted and hopefully stands a better
chance of not having any unwanted bugs.

This new version supports the `search_contains` option which I want to
use so that searching for multiple words will work.