Fix invisible tab
Bug description: When viewing a policy, the history and notes tab on the right
had a count, but you couldn't see the data unless you selected the other tab
and then came back to this tab.

It's essential that the first pane in a set of tabbed content has the class
"active", or it will not be shown when the page loads. This missing class was
the problem here.

It's better if the programmer doesn't have to worry about forgetting things,
so by delegating generation of the tab wrapper to an object that encapsulates
the is-first-tab? state, we can ensure that the first tab is always active on
page load.

As a nice side effect, this also makes tab rendering terser, and eliminates
some spaghetti templating where which tab was first depended on some
conditional logic.

I'm not entirely sure that TabPaneState is the best name for this class, but
that's not a pressing concern.