Add Military roles.
This is to allow us to model the Chiefs of staff for the MOD.  I spent
some time trying to move away from STI, and toward a RoleType equivalent
of SpeechType.  Unfortunately I ended up going round in circles and
decided to follow the pattern of what's there for now.

The main changes in this commit are to:

* Display the current Chief of the Defence Staff on
  /organisations/<org-name>/

* Add /organisations/<org-name>/chiefs-of-staff

A future commit will add the ability to create Military roles.