Display the correct role type when editing roles.
I noticed that the 'Type' selectbox would always display Cabinet
minister when editing a Role, irrespective of that Role's type.

The problem was that we cast a specific role instance (e.g. BoardMemberRole)
into a Role in the _form partial. This meant that
RoleTypePresenter.option_value_for couldn't determine the currently
selected type and would default to cabinet_minister.

I'm not entirely sure about this fix (storing the role_type before
casting to a Role) but it does the trick for now.  I thought about
introducing a specific test to ensure we're selecting the correct option
but wasn't convinced the overhead was worth it.  I ended up modifying
the "should display form for updating an existing role" test instead.