First pass at paginating specialist guides on the client.
I've implemented two generic javascript mechanisms - one to split
the children of an HTML container into "page" divs based on a certain
tag delimiter (like 'H2'), and another to produce a list of links to
headings within a container.

On this, I've build the simple javascript required to paginate a
document on the client side by splitting guides on H2 elements into
pages, adding sub-page navigation to the existing document section
navigation, and then hiding all the pages except the first. Writers
should use H3 elements for internal page navigation.

There are still some unresolved issues with this:

* H4 elements appear unstyled at the moment;
* Footnotes still appear at the end of the document, which is the
  last page once pagination has occured, but it would probably be
  better to pull footnotes onto the same page they are referenced
* Govspeak markdown doesn't allow continuing numbered lists (there is
  already a story for this:
  https://www.pivotaltracker.com/story/show/31442843)

Testing is also a question; since this depends quite heavily on some
of the markup that's present when a page is rendered (i.e. where the
summary, content, and document navigation is) it feels like the best
way to test would be using the rendered page output, rather than a
javascript fixture. I'm still considering how practical it would be to
add a proper javascript capybara web driver.