Allow attachments on specialist guides.
While adding this was pretty simple, it did reveal a few design
quirks in our controller code. Specifically, ensuring that the
attachments were built properly under error conditions was only
working in most cases because the NationalApplicability controller
mixin was doing the work. The fact that there was an empty
`build_edition_attachments` method in that module should've given this
away.

I've extracted the relevant functionality into a method that can be
overridden by subclasses as a start to unpicking the two features.