The Jenkins build should fail if production asset precompilation fails.
A comment block around commented out lines in a SCSS file causes asset
precompilation to fail. When this happened in
c4fee13facd26839130f6936aa9e8fc529681501 this had the unfortunate
consequence that after a preview deployment, the server was left with no
`current` directory and therefore the app was responding with a 403
Forbidden.

In order to avoid bringing the preview environment down when
precompilation fails on an automated preview deployment, we can make
the build fail to highlight the problem as soon as possible and
hopefully make it easier to diagnose.