Add publication subtypes.
PublicationType is a non-persisted psuedo-active-record model class.

The 'Unknown' publication type is used as a default value when
migrating. Existing documents will need to be reviewed manually to set
the real publication type.

I refactored the publication document_contextual partial to accept an
edition as a local var rather than referring to the global @document,
allowing it to be used in the admin preview.