The Jenkins build should fail if production asset pre-compilation fails.
This is a re-attempt of 3c594f41c92de7dc1826041f389b5b1d599b0a6b, but
this time creating an empty production database (using the `db:create`
rake task) in order to keep the `assets:precompile` rake task happy when
it loads the Rails production environment which in turn accesses the
database (e.g. when observers are loaded). It's not ideal that we are
creating a "production" database on the continuous integration server,
but it is at least only an empty database, and I couldn't find any other
non-hacky way around this.