The Jenkins build should fail if production asset pre-compilation fails.
This is another re-attempt of 3c594f41c92de7dc1826041f389b5b1d599b0a6b,
this time specifically skipping the loading of observers when loading
the Rails environment for the `assets:precompile` rake task. This means
we no longer have any need for a production database on the Jenkins box.

It turned out that in the last re-attempt
(d019529416bf9cf1fea8499c4076d06bf15f0ab8), we didn't have permission to
create a new database on the Jenkins box.

As explained in the original commit note the reason to make this part of
the CI build is to catch problems with pre-compilation in the Whitehall
build rather than waiting for the deployment of Whitehall to preview to
fail spectacularly giving 403 Forbiddens.