Improve signal-to-noise ratio in CI build console output.
The default Cucumber formatter outputs all the steps in all the
features as well as reporting errors and failures. When a build fails it
takes a while to look through all the output to find the error/failure.
Using the "progress" formatter gives much terser output, but retains
all the useful info for errors & failures.