Run Cucumber features as part of the CI build.
We had accidentally stopped doing this in
9f5c67e5512eb24a7dfa00888ce5be4797de4269, when the `test:cleanup` task
was added. I think what happened was that when the rake command was
changed to run multiple tasks (i.e. because we also wanted to run
`test:cleanup`), we needed to be explicit about running all the normal
test tasks. However, instead of using the `default` rake task, we
chose to use the `test` task. Unfortunately the `cucumber` task is only
a dependency of the `default` task and not the `test` task.

This happened about a month ago, but happily, because developers tend to
use the `default` rake task on their local machines and are diligent
in running the tests locally before committing, all the cucumber
features still pass. Yay!