Remove links to individual supporting documents.
We've decided to replace the links to individual documents, on the
policy page, with a single link to "Detail" which takes you to the
first supporting document page.  I've done this by linking to
/supporting-pages and then having that action redirect to the first
page.  I'm also, possibly controversially, rendering a 404 if you
request /supporting-pages and there aren't any documents to display.