Group specialist guides by topic
Added tests to specialist_guides_controller_test to check the headings.
We made the html output include a div to wrap each topic to facilitate
testing.

We also added a new scope to topic to fetch published specialist guides
which allows us to include them in the Topic query more easily.

Finally, we made some amends to the cucumber tests to reflect the fact
that specialist guides will only appear on the specialist guide index
page if they have been assigned to a topic. This behaviour was
explicitly requested by Neil on the basis that they want to require that
specialist guides have a topic set. That validation hasn't been
implemented yet.