Avoid a javascript warning in specialist guide javascript unit tests.
We were seeing the following warning on the CI machine which is running
v1.3.0 of phantomjs :-

    TypeError: Result of expression 'newPosition' [null] is not an object.

I have v1.5.0 installed locally and I don't see this warning.

There isn't an official package for v1.5.0 on Ubuntu Lucid and
so given this is just a warning, it didn't seem worth spending a long
time working out how to upgrade phantomjs.