Reduce duplication in document_related partial.
We had two identical 'Topics' lists in this partial.  One that worked
depending on whether the supplied document had topics (news articles,
policies, supporting pages and specialist guides) and one that worked if
you passed topics in to the partial (consultation responses,
consultations, publications and speeches).  We're now always passing the
topics in to the partial.

*NOTE* I had to add a test to specialist_guides_controller_test that
checked that we were displaying the topics.  Without this, this commit
would've resulted in us no longer displaying the topics for a specialist
guide.