Re-order edition columns to avoid discrepancy in schema.rb.
Two migrations (20120628123047 & 20120619153706) were committed in
reverse timestamp order. This meant that if you create a database from
scratch using the migrations or from loading the latest db/schema.rb
file in the git repository, you ended up with a different schema from
the one in preview and production. This makes things awkward when people
import data from preview or production and re-generate the schema
locally (e.g. when running tests), because it puts the columns in a
different order in schema.rb and shows up as a difference in git.