Clean assets before running tests.
We've been seeing intermittent failures in our Specialist Guide
javascript tests when running under Jenkins.  I think the problem is
being caused by the pre-compiled assets hanging around from a previous
run.

This is the sequence of events that I think explains the problem
we're seeing.

* Code pushed to Jenkins, all tests pass and the assets are compiled.
* A change is made to the assets, the code is pushed and a javascript
  test fails because it's using the old assets.  Importantly - the new
  assets aren't compiled because that only happens if all the tests
  pass successfully.
* We comment out the 'failing' test, push to Jenkins, all tests pass and
  the new assets are compiled.
* We re-enable the test and see it pass on Jenkins.
* As soon as we make another change to the assets we're right back at
  step 1.

So, cleaning the precompiled assets before running the tests should mean
that we're always using the most recent assets that we're expecting.

*NOTE* I've renamed PRECOMPILING_ASSETS to the more generic
SKIP_OBSERVERS_FOR_ASSET_TASKS as it's now being used when precompiling
*and* cleaning assets.