Automatically add rel=external to external links in govspeak.
Use the functionality added in 74ea8c77293ffa2ef7c2876e1b832072ae297017
to automatically add the rel attribute to external links in a more
robust way.

This means it is now possible to add an external link in the middle of a
paragraph. The old-style 'x' delimited link syntax didn't work in this
situation.

Note that the old syntax is still allowed, so we don't break any
existing content, but we should prefer not to use it. I plan to update
the govspeak help to encourage this approach.

Note also that we have to supply 2 document domains to govspeak, because
of the wacky URL re-writing of admin URLs to public-facing URLs.