Slimmer was not logging to the Rails log as intended.
Setting the Slimmer logger to be the Rails logger doesn't work when
within the Application initialization, because the Rails.logger is `nil`
at this point which means Slimmer ends up using its default
`NullLogger`.

We can fix this problem by moving the configuration of the Slimmer
logger into a Rails initializer file at which point the Rails logger has
been set.

The original configuration was set in
9ec2ab1516981c2b00dd4248ec5e0b20d89f3a64.