Make Slimmer cache templates.
I was investigating why we were seeing some exceptions due to Slimmer
requesting templates from Static when the Static app was down e.g.
during deployment of Static. @heathd pointed out that Slimmer should
cache these templates in memory and so it should only be when the
Whitehall app has just been restarted that we should see such errors. I
decided to investigate further to see whether Slimmer was configured to
do this within the Whitehall app and discovered that it was not.

We want Slimmer to cache templates in production, because they don't
change very often and each page requires quite a few of them. Doing a
network round trip for each of these templates must be adding quite a
bit to our page load times, especially as Static is on a different box
to Whitehall.