Tidied up indentation and whitespace.
Up until now I've been trying to preserve the (horrible) state of these
Nomensa Accessible Media Player files in order to make it easier to diff
against newer versions of the code. But given this file isn't a core
part of the media player library and I've already modified it
considerably, I decided it was time to tidy it up.