Make javascript a little more idiomatic.
Instead of using the `js-fixed-content-top` CSS class to activate the
javascript behaviour, we declaratively apply the behaviour in
`on_ready.js`. This seems to fit in better with the approach we've taken
with other javascript functionality and it allowed me to reduce the
duplication of the use of `$(".js-fixed-content-top")` which appeared
multiple times in the old implementation.