Style updates to Specialist Guidance
Refactored js-nav-helper to use specific `js-` prefixed class. This let
us easily remove the helper from the specialist guidance pages.

Added some @media styles to make the pages work better when displayed on
a small screen.

Updated the colour of document headings to be a grey as we are no longer
using NTA-Bold variant on the site and some differentiation was needed.