Make Organisation links on Guidance pages external
Extended the organisations_logo_list partial to be able to use external
organisation urls and to not add line breaks into the names.

Added a test to ensure that the organisation links on Guidance pages are
to the external sites

Removed the now unused helper to turn a list of organisation into a
sentence. Also removed the tests associated to the helper.