Fix a bug in the conflict partial template.
We were checking whether an edition could be associated with related
policies but displaying topics, and checking whether an edition could
be associated with topics but displaying related policies.

This caused an exception to be raised when the conflict resolution
template was displayed for a specialist guide, as a guide can be
associated with topics, but *not* with related policies.

There were existing tests, checking for a specific heading within the
conflict resolution container, that I was able to extend to check
for the actual related policies and topics.  These gave me failing tests
that I could use to determine that the bug was fixed.

For reference only, the problem was introduced in a combination of
97437877d47475e357680c1c4c3251c050134709 and
73978dd9c4f9eb207e3dc97d58f9ebac264d7cf4.