Set Google Analytics custom variable `Proposition`.
To be consistent with the mainstream apps, we should be setting this
custom variable to a value dependent on whether the page is showing
specialist guidance related content or inside government content.

Calling Slimmer::Headers#set_slimmer_headers adds headers to the
response which is then interpreted by the Slimmer middleware and the
custom variable assignment javascript is written into the relevant
script tag in the final response.

It would be better to test this in an integration test that actually
uses Slimmer for real, but this is better than nothing.