Group guidance by either topic or organisation
Added specialist_guide association to organisations and
published_specialist_guide. Added a unit test to ensure that
published_specialist_guides returns the correct guides.

Added the ability to group by organisation in the
specialist_guides_controller.

Refacotred specialist_guides#show view to be agnostic about the grouping
of speclialist guides and added navigation to switch to organisation
grouping. Added test to ensure that the flag worked as expected.