Add Command paper number field to Publications.
I've added some basic validation as per Neil's request in the story,
although I do wonder whether we'll want to relax it a little.  I've
already spotted a couple of things that we might want to handle:

1. TNA[1] lists 'C' as a valid prefix, while Wikipedia[2] has it as 'C.'
As it stands, only 'C.' is valid.
2. Wikipedia lists a set of numbers (1 to 4222) that are valid without
any prefix, while TNA suggests that a prefix is always required.

[1]
http://www.nationalarchives.gov.uk/information-management/our-services/command-papers.htm
[2] http://en.wikipedia.org/wiki/Command_paper