Move the Publication type dropdown up the publication page.
This is a required field so having it so far down the form was just
annoying.

Because of our use of the standard_edition_form helper and template
inheritance in the admin interface, I've ended up adding another
partial ('additional_required_fields') that can be used to inject form
elements under the body textarea.