Remove the site/headers debug action to prevent slimmer conflicts.
The site/headers action isn't being used, but having a method in our
controller called headers conflicts with setting the slimmer
proposition header (as it does this using a #headers method).  The
site controller was originally intended to just show useful (but
harmless) information for developers, such as the sha of the current
code.  It has since evoled to serve other pages including the sunset
page, leading to this conflict.  It would probably be better to
split these responsbilities between two controllers.