Split responsibility of SiteController
HomeController is now responsible for 'homepage' like things, such
including the home page, the sunset page and the site tour.  These
are all things that general visitors will be interested in.

SiteController is only now responsible for actions of most interest
to developers, such as the site sha and the grid (whatever that is).