Update plek.
Plek has been changed so that it no longer discards underscore
characters in service names.  For example 'whitehall-search' and
'whitehallsearch' were previously considered as the same service,
but are now seen as different. Updating the gem now to ensure we
don't ever start to rely on the previous broken behaviour.