Remove audit trail entries accidentally introduced previously.
Previous migrations that have updated Edition have accidentally
created audit trail entries.  The migrations that have done
this are:

    20120608123237_remove_non_ssl_edition_video_urls.rb
    20120611092053_remove_ssl_edition_video_urls.rb
    20120719081325_make_attachments_inline.rb

These audit trail entries are broken as they don't have a whodunnit
attribute set, causing problems when displaying these editions.

It would be great to add a null constraint to prevent this from
happening again, but doing so breaks a ton of tests so will be
done separately.