Change notes clearer and more accessible
If there are no changes to a document only the published date needs to
be shown. If there are changes then show the update notes.

Updated toggle_change_notes.js to no longer use a link to toggle the
change notes. This was a result of an accessiblity check from Léonie.
Screen readers will now just see it as a heading followed by some text
without any interaction.

Updated some tests which were wrongly looking for an updated time on the
page. We should only display published dates or dates which have change
notes.