Use first_published_at as first published_at
In the change notes the first published_at should be the date of the
first_published_at. The first_published_at is the date the document was
first published possibly on a previous system to GOV.UK and should be
used as such. The published_at date only has meaning for edits while in
the GOV.UK system.