Get the ajax cucumber feature working.
It turns out that this feature was behaving correctly, and two
publications really were being displayed. This is because the
'All Departments' option was still checked when the form was
submitted.

Trying to unselect this option revealed a bug in capybara-webkit,
which does not recognise `<select multiple ... >` as a multi-select
tag. [I've submitted a fix to the upstream repository][1].

I wasn't completely confident in my fix -- this is my first commit
to capybara-webkit, and should be checked -- so rather than use the
patched library, I worked around the bug by implementing the
deselection via javascript driven by capybara. As soon as a fix is
approved upstream, this hack should be removed.

I also took the liberty to move from using "Fill in"-style steps to
higher-level descriptive steps, although this could certainly have
been done separately.

[1]: https://github.com/thoughtbot/capybara-webkit/pull/361