Make Guidance importer take slugs instead of names
There's a bug in Rake where commas in arguments are treated as argument
separators, even when quoted. This means departments like "Department
for Environment, Food & Rural Affairs" don't work at all.

This changes the arguments to use the slug for "topic" and
"organisation" instead, which will be unambiguous due to slug format
rules.

Also a general tidy up of error messages and desc to make it more `rake
-T` friendly.