Fix test failure due to test interference
The test interference is due to some users existing in the test database
which were created by previous tests. The existence of users in the
database allows the 'mock' authentication strategy to succeed (which
logs in the first user in the DB).

This is a hacky fix, I think that the correct fix would be to change the
logic around GDS_SSO_MOCK_INVALID in the GDS-SSO gem. However, I'm not
confident about the intention/usage of this flag.