Add ISBN field to publication attachments in admin.
*NOTE* There is still duplication between the ISBN field on a
Publication (Edition) and the new ISBN field on an attachment.  This'll
be removed in a subsequent commit.

The 'attachments' partial is used for Consultations, Consultation
responses, Publications, Specialist guides and Supporting pages.  We
only want the ISBN field to appear for publications so I've amended the
partial to render the 'additional_attachment_fields' partial which is
empty for everything except Publications.