Copy Price from Publication to Attachment.
It's possible for each of a publication's attachments to represent a
physical document that can be purchased.  Moving the price from the
Publication object to the Attachment means that we're now modelling this
correctly in the system.

This is the final attribute to move from Publication to Attachment.  The
others were the ISBN, Unique Reference number, Command paper number and
Order URL.