Fix bug in multiple file upload javascript.
The bug caused the new file upload container ('.file_upload') to be
copied right below the file field in the case where the file field was
already surrounded by the 'field_with_errors' container.  I was able to
add a test to expose the buggy behaviour and subsequently fix it.