Show first published as first change note
The date at which a document is added to inside-government is only
important if there is no origianl publication date. In that situation
the original date should be used.

This should make the interface less confusing as there is only one
publication date for a publication or document. All changes are then
listed after this date.