Replace hand-rolled pagination with an external library.
This should reduce our testing burden. Kaminari is also used in
Publisher for pagination, and appears to be under reasonably active
development.

One nice aspect of using this gem is that we can re-implement the
existing pagination style by overriding the default partials from
the gem, as I have done here.