Update migration to use SQL rather than AR models.
This migration currently fails in production because it tries to update
the Attachment model, which in turn tries to store data in the
price_in_pence column that doesn't yet exist (it's introduced in a later
migration).  I did try to redefine the AR models in the migration but I
just couldn't get it to use my new Attachment model: It continued to try
to use the class in app/models.

I diff'd the debug output of the current and previous implementation of
this migration to assure myself that they were doing the same thing.