Enable dynamic filtering on the specialist guides page.
The bulk of this work was making the (yet to be renamed) javascript
generic enough to work on multiple document types. Notable parts of
this effort:

- making classes/IDs in the markup more generic for containers, and
  more specific for attributes (date -> publication_date)
- extracting publication-specific styles into the publications CSS
  files, to clarify what the core of the document filtering styles
  are
- Changing the javascript to dynamically determine the attributes
  for display in the table, ignoring core attributes like id and
  title
- Extracting some of the common filtering cucumber steps into their
  own definitions file

Next, I am going to rename the javascript file (I didn't in this
commit for clarity of diff), add some controller tests for JSON
output, and ensure that the topics shown have relevant content.