Update document_page ready for new designs.
The new designs have moved a whole collection of data around and this
should be reflected on the master document_view template rather than
only working on each pages template.

Moved change-notes into top right corner and update styling to account
for its new position.

Moved country, topic, inapplicable nation information up to the top
right of the document and unify the genrators for the paragraphs.

Updated the tests for reflect the new structure changes.