Add another link to mainstream content.
This is basically a copy of 6df161c197b5c579833285161b93d2d5419e49a9.

It was found that linking to a single mainstream guide isn't enough for
some of the specialist guidance we've created.  This solution isn't
elegant but it's the quickest thing we can do given that we want it
ready for the launch of specialist guides.

I presume that we'll want the ability to link to multiple mainstream
guides in the near future.  As soon as that becomes a requirement we'll
want to replace the current implementation with something more flexible.