Add brakeman to our Gemfile.
We run brakeman as part of the build process on Jenkins.  Version 1.3
(which is installed on ci.alphagov.co.uk) takes 15 seconds to run on my
local machine.  The version I've just added to the bundle (1.7.0) takes
around 6.5 seconds on my machine.  Builds on Jenkins take a long time to
run so anything we can do to make them slightly faster seems like a Good
Thing.

I've updated the Jenkins configuration to use the bundled version of
brakeman and will monitor the output for any speed improvements.