Ensure govspeak markdown extensions are rendered in atom feeds.
In order to do this, I had to make sure that the attachments
partial was rendered using the 'html' format; it seems by default
Rails wants a `document/_attachment.atom.erb` partial to exist.

It's worth noting that explicitly setting the filename like this
may have consequences later on, particularly if the application
templates are ever loaded based on locale. I believe that Rails 3.2
might allow us to pass a `formats: [:html]` parameter to the
`render :partial` call, but we're not in a position to upgrade
right now.