Add a test for eager loading of *all* associated organisations.
We could also add more of an integration test around the filtering,
since that's what exposed the bug.

It's worth noting that this bug exists because of the combination of
`includes(:organisations)` from the controller, and `where(:...)` in
the `in_organisation` method. This test attempts to assert that both
eager loading and organisation filtering are both operational.