Don't include date filters in atom feed URLs.
It doesn't make sense to have an atom feed which will only ever
contain entries about documents from a specific date or earlier.

I had hoped to keep the atom feed URL out of the JSON response, but
I felt that the duplication of this logic was worse than including it; stripping out the date parameters is not appropriate for the URL
of the actual JSON data.