Use background image for external links
Using an `:after` selector with a url content was causing issues in IE8.
This is fixed by using a background image instead.

The specific issue it was causing was the image wasn't shown if the
element was hidden using javascript at some point. As we are using
javascript to page specialist guidance any external link on the
non-landing page would just have a strange space after it.