Update kramdown to git bugfix for abbr within mailto
There were two issues fixed in kramdown master which have not yet been
included in a published gem. The fixes should be in the next version
which would be 0.13.8, but until that's released, we're pointing to a
github url.

Unfortunately the kramdown repo doesn't include a separate gemspec, so
we created a fork which includes a gemspec.

The fixes we needed are:

* https://github.com/gettalong/kramdown/commit/0c7ee71cde6eaf67b76ed048fda0c2dbdbd6480f
* https://github.com/gettalong/kramdown/pull/13