Consistently test attachment behaviour of editions.
Our attachment tests were kinda spread all over the place.  We had some
in attachable_test, some in publication_test and some in
consultation_test.  I've chosen to move the tests for the attachable
behaviour into attachable_test and then just test that we're including
the Attachable module in the relevant document tests.  This isn't a
foolproof method of testing but it feels like a good compromise between
comprehensive coverage and speed.