Allow toggling of inline attachment capability
This is in preparation for disabling the inline attachment help sections
in the admin interface.  This'll be for Publications initially as the
new template contains a link to a single 'primary' attachment and then a
list of all attachments at the bottom of the page.

*NOTE* There's a separate story to completely prevent the use of inline
attachments, although I imagine that'll also use these flags.

*NOTE* I've made the `edition_class_from_test_name` method available to
all `ActiveSupport::TestCase`s so that I can make the test helper
methods slightly simpler.