Extract tests for showing inline attachments help
I've extracted the test that ensured we were displaying the markdown
help in the form and I've added two new tests to ensure we're showing
the govspeak help in the sidebar.  My hope is that this'll make it
easier to reverse the test for Publications now that we're no longer
supporting inline attachments.