Prevent NewsArticles and Speeches from having the same slug
Since both document types will be moved to live under
/government/news-and-speeches we'll need to ensure that their slugs
don't clash.

It turns out that friendly_id isn't very friendly in this respect. The
'scoped' plugin of friendly_id only offers the option of scoping
potential conflicts by the value in a particular column, or the value
held in a related table through an association.

This was the least ugly way I could find to treat NewsArticle and Speech
as equivalent when checking for slug conflicts. It relies on
overriding the private method FriendlyId::SlugGenerator#conflicts. This
is also what the 'scoped' plugin does, but since that's in the same repo
it's obviously less sucseptible to breakages over time.