Fixed test which caused phantomjs timeouts
Previously this test was submitting a form to take you to a new page
which didn't have QUnit on it. PhantomJS was then looking for something
which didn't exist and that was causing a timeout.

By mocking pushState the url shouldn't change and leaves the test page
in a better situation for future tests.