Fix escaping of organisations on speech page
The original behaviour worked fine when there was only a single
organisation but calling `#to_sentence` on an array of organisations
left us with html escaped anchor tags, which isn't what we wanted.  I've
added a failing test and fixed the implementation by marking the
resulting string as `#html_safe`.